requirements.txt
autopep8==1.3.4
Django==2.0.3
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
django-webpack-loader==0.6.0
pycodestyle==2.3.1
pytz==2018.3
